[Gradient Compression] Add an index field to GradBucket for PowerSGD
Pull Request resolved: https://github.com/pytorch/pytorch/pull/48757

Add an index field to GradBucekt, so error_dict is keyed by this index instead of the hashcode of input tensor. The replacement will be done in a separate diff, as the definition of this new method somehow couldn't be recognized in the OSS version.

Original PR issue: Investigate Applying PowerSGD to Communication Hook for Gradient Compression #47202
ghstack-source-id: 117916685

Differential Revision: [D25288496](https://our.internmc.facebook.com/intern/diff/D25288496/)